Improve feedback while waiting for CFN stacks to be created

+++ workflows/environment_terminate.go
@@ -1,7 +1,9 @@
 package workflows
 
 import (
+	"fmt"
 	"github.com/stelligent/mu/common"
+	"strings"
 )
 
 // NewEnvironmentTerminator create a new workflow for terminating an environment
@@ -52,7 +54,11 @@ func (workflow *environmentWorkflow) environmentEcsTerminator(environmentName st
 			return err
 		}
 
-		stackWaiter.AwaitFinalStatus(envStackName)
+		stack := stackWaiter.AwaitFinalStatus(envStackName)
+		if stack != nil && !strings.HasSuffix(stack.Status, "_COMPLETE") {
+			return fmt.Errorf("Ended in failed status %s %s", stack.Status, stack.StatusReason)
+		}
+
 		return nil
 	}
 }
@@ -65,7 +71,10 @@ func (workflow *environmentWorkflow) environmentVpcTerminator(environmentName st
 			log.Debugf("Unable to delete VPC, but ignoring error: %v", err)
 		}
 
-		stackWaiter.AwaitFinalStatus(vpcStackName)
+		stack := stackWaiter.AwaitFinalStatus(vpcStackName)
+		if stack != nil && !strings.HasSuffix(stack.Status, "_COMPLETE") {
+			return fmt.Errorf("Ended in failed status %s %s", stack.Status, stack.StatusReason)
+		}
 
 		targetStackName := common.CreateStackName(common.StackTypeTarget, environmentName)
 		err = stackDeleter.DeleteStack(targetStackName)
@@ -73,7 +82,11 @@ func (workflow *environmentWorkflow) environmentVpcTerminator(environmentName st
 			log.Debugf("Unable to delete VPC target, but ignoring error: %v", err)
 		}
 
-		stackWaiter.AwaitFinalStatus(targetStackName)
+		stack = stackWaiter.AwaitFinalStatus(targetStackName)
+		if stack != nil && !strings.HasSuffix(stack.Status, "_COMPLETE") {
+			return fmt.Errorf("Ended in failed status %s %s", stack.Status, stack.StatusReason)
+		}
+
 		return nil
 	}
 }
